Refactor: 벌크 연산을 적용해 UserLoginHistory를 삭제해라 - #105

+++ tictoc-batch/src/main/java/tictoc/userLoginHistory/config/UserLoginHistoryBatchConfig.java
@@ -14,10 +14,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.transaction.PlatformTransactionManager;
 import javax.sql.DataSource;
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @Configuration
@@ -27,7 +26,6 @@ public class UserLoginHistoryBatchConfig {
     private final DataSource dataSource;
     private final PlatformTransactionManager transactionManager;
 
-
     @Bean
     public Job userLoginHistoryJob(Step step) {
         return new JobBuilder("userLoginHistoryJob", jobRepository)
@@ -35,11 +33,10 @@ public class UserLoginHistoryBatchConfig {
                 .build();
     }
 
-
     @Bean
     public Step userLoginHistoryStep() {
         return new StepBuilder("userLoginHistoryStep", jobRepository)
-                .<Long, List<Long>>chunk(100, transactionManager)
+                .<Long, Long>chunk(100, transactionManager)
                 .reader(userLoginHistoryReader())
                 .writer(userLoginHistoryItemWriter())
                 .build();
@@ -47,22 +44,23 @@ public class UserLoginHistoryBatchConfig {
 
     @Bean
     public JdbcCursorItemReader<Long> userLoginHistoryReader() {
-        LocalDateTime startDate = LocalDateTime.now().minus(1, ChronoUnit.YEARS).withMonth(2).withDayOfMonth(28);
         LocalDateTime endDate = LocalDateTime.now();
+        LocalDateTime startDate = endDate.minusDays(7);
 
         return new JdbcCursorItemReaderBuilder<Long>()
                 .dataSource(dataSource)
                 .name("userLoginHistoryReader")
                 .sql("SELECT id FROM user_login_history WHERE login_at BETWEEN ? AND ?")
-                .queryArguments(startDate, endDate)
+                .queryArguments(Timestamp.valueOf(startDate), Timestamp.valueOf(endDate))
                 .rowMapper((rs, rowNum) -> rs.getLong("id"))
                 .build();
     }
 
     @Bean
-    public ItemWriter<List<Long>> userLoginHistoryItemWriter() {
+    public ItemWriter<Long> userLoginHistoryItemWriter() {
         return items -> {
             NamedParameterJdbcTemplate namedParameterJdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
+            if (items.isEmpty()) return;
             String sql = "DELETE FROM user_login_history WHERE id IN (:ids)";
             Map<String, Object> params = new HashMap<>();
             params.put("ids", items);
